use hook to reduce `CadView` size
